Moved environment to a header file
